models.py -> loaders/ module refactor
